# Reduce allocations/copies in `String::format`